Feat/be-profileimage (#94)
* [BE feat] #72 회원 프로필 이미지 - s3 버킷 연동
1. 프로필 이미지 관리 장소 변경(local -> AWS s3 Bucket/ProfileImages)
   - 스토리에 업로드 되는 파일은 's3 Bucket/StoryFiles' 폴더에서 하시면 될 것 같습니다.
2. AWS Configuration 설정 추가
3. MatchBoardDto.Response 필드 추가 (프로필 이미지 링크)
   - MatchBoardMapper 수정

4. application.yml 파일 수정했으나, 민감한 정보가 있어서 같이 커밋하지는 않겠습니다. 월요일에 yml 설정 방법 간단히 말씀드릴 예정입니다.

* [BE refactor] #41 회원 프로필 이미지 CRUD
1. MatchBoardService.java 매칭하기 수정, 삭제 권한 확인 로직 추가
2. s3 defaultImage.png 삭제
   - 기본 프로필은 db만 저장, 사진 파일은 저장 x
3. SimpleLoggingAop.java 포인트컷 추가
   - inSecurity()
4. updateProfile (PATCH) 수정
   - 'image' 데이터가 활성화(Postman 기준 체크 표시 on) & 업로드 파일이 없다 -> 프로필 이미지 삭제
   - 'image' 데이터가 활성화(Postman 기준 체크 표시 on) & 업로드 파일이 있다 -> 프로필 이미지 업로드
   - 'image' 데이터가 활성화 x(Postman 기준 체크 표시 off) -> 아무 일도 일어나지 않는다
   - 프로필 이미지 content-type 검증(only image)
5. 회원탈퇴(deleteMember) 시 프로필 이미지 파일 삭제 (from s3 bucket)
6. authController.java -> AuthController.java 이름 변경

+++ server/src/main/java/mainproject33/domain/member/service/ProfileImageService.java
@@ -7,32 +7,25 @@ import lombok.extern.slf4j.Slf4j;
 import mainproject33.domain.member.entity.Profile;
 import mainproject33.domain.member.entity.ProfileImage;
 import mainproject33.domain.member.repository.ProfileImageRepository;
+import mainproject33.global.exception.BusinessLogicException;
 import mainproject33.global.exception.ExceptionMessage;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 import java.util.UUID;
 @Service
 @Slf4j
 @RequiredArgsConstructor
 public class ProfileImageService {
     private final ProfileImageRepository imageRepository;
-    private final AmazonS3 amazonS3;
     private final String bucket = "gameto";
-    private final String defaultFileName = "defaultImage";
-    private final String defaultExtension = ".png";
+    private final AmazonS3 amazonS3;
 
     public ProfileImage createDefaultProfileImage(Profile profile) { // 기본 프로필 이미지 생성
-        String uploadFileName = defaultFileName + defaultExtension;
-        String storeFileName = UUID.randomUUID() + defaultExtension;
 
-        saveDefaultImageFile(storeFileName); // 이미지 파일 로컬에 저장
-
-        ProfileImage profileImage = saveImageData(new ProfileImage(), uploadFileName, storeFileName); // 이미지 데이터 DB 에 저장
+        ProfileImage profileImage = saveImageData(new ProfileImage(), null, null); // 이미지 데이터 DB 에 저장
         profile.setImage(profileImage);
 
         return imageRepository.save(profileImage);
@@ -40,28 +33,30 @@ public class ProfileImageService {
 
     public ProfileImage updateProfileImage(Profile findProfile, MultipartFile file) { // 프로필 이미지 수정
         deleteImageFile(findProfile.getId()); // 기존 이미지 파일 삭제
+        ProfileImage profileImage;
 
-        String uploadFileName = file.getOriginalFilename();
-        String storeFileName = UUID.randomUUID() + parseFileExtension(file);
+        if (file != null && !file.isEmpty()) {
+            String uploadFileName = file.getOriginalFilename();
+            String storeFileName = UUID.randomUUID() + parseFileExtension(file);
 
-        saveImageFile(storeFileName, file); // 이미지 파일 로컬에 저장
+            saveImageFile(storeFileName, file); // 이미지 파일 로컬에 저장
 
-        ProfileImage profileImage = saveImageData(
-                findProfile.getImage(), uploadFileName, storeFileName); // 이미지 데이터 DB 에 저장
+            profileImage = saveImageData(
+                    findProfile.getImage(), uploadFileName, storeFileName); // 이미지 데이터 DB 에 저장
+        } else {
+            profileImage = saveImageData(
+                    findProfile.getImage(), null, null);
+        }
 
         return imageRepository.save(profileImage);
     }
 
     public String readProfileImagePath(Long id) { // 프로필 이미지 가져오기
-        ProfileImage image = imageRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException(ExceptionMessage.IMAGE_DATA_NOT_FOUND.get()));
+        ProfileImage findImage = findVerifiedImage(id);
+        String storeFileName = findImage.getStoreFileName();
 
-        return amazonS3.getUrl(bucket, "ProfileImages/" + image.getStoreFileName()).toString();
-    }
-
-    private void saveDefaultImageFile(String storeFileName) { // 기본 프로필 이미지 파일 저장
-        amazonS3.copyObject(
-                bucket, "ProfileImages/defaultImage.png", bucket, "ProfileImages/" + storeFileName);
+        if (storeFileName == null) return null;
+        else return amazonS3.getUrl(bucket, "ProfileImages/" + storeFileName).toString();
     }
 
     private void saveImageFile(String storeFileName, MultipartFile file) {
@@ -82,16 +77,17 @@ public class ProfileImageService {
         return profileImage;
     }
 
-    private void deleteImageFile(Long memberId) {
+    public void deleteImageFile(Long memberId) {
         ProfileImage findImage = findVerifiedImage(memberId);
-        amazonS3.deleteObject(bucket, "ProfileImages/" + findImage.getStoreFileName());
+
+        if (findImage.getStoreFileName() != null) {
+            amazonS3.deleteObject(bucket, "ProfileImages/" + findImage.getStoreFileName());
+        }
     }
 
     public ProfileImage findVerifiedImage(Long memberId) {
-        Optional<ProfileImage> optionalImage = imageRepository.findById(memberId);
-        ProfileImage findImage = optionalImage.orElseThrow(() -> new RuntimeException("프로필 이미지를 찾을 수 없습니다."));
-
-        return findImage;
+        return imageRepository.findById(memberId).orElseThrow(() ->
+                new BusinessLogicException(ExceptionMessage.IMAGE_DATA_NOT_FOUND));
     }
 
     private String parseFileExtension(MultipartFile file) { // base64EncodedFile 확장자 Parser. TODO : 추가 확장자 확인 필요
